Added the code for recognizing non-root nodes to netfs_S_file_get_translator_cntl
Now netfs_file_get_translator_cntl will recognize whether it
is called for the root node or not. In case the node it is
called on is not the root, it will call file_get_translator_cntl
on the real node, so that filters could trace translator stacks
on real nodes.