Added the port to the untranslated version of the real node to the netnode
When 'file,,x' is requested, nsmux cannot tell whether 'x'
is a filter or not and whether it will require the untranslated
version of the file or not. However, when
netfs_S_file_get_translator_cntl would like to show the
translator stack on the *real* node, it will need the
untranslated version of the node, which must be available by
that time. The new field in netnode solves this issue.