Magic syntax handling removed from netfs_attempt_lookup_improved
Up to now it was netfs_attempt_lookup_improved who handled the
magic syntax for ``,,''. Now the idea is to handle this syntax
via retries and netfs_attempt_lookup_improved is in no position
to manage this. In further commits the magic syntax will be
handled by netfs_S_dir_lookup. Note nsmux does not actually
work at the moment.