Added the function to create proxy nodes for ports
Since we want netfs_S_dir_lookup to ask the client to do a
retry when it sets a new dynamic translator, nsmux has to give
the client a *proxy* of the port to the root of the current
dynamic translator, so that the retry gets back in nsmux. For
this we need to create proxy nodes for ports, which is precisely
what node_create_from_port does.