Made nsmux able to pass retries to the client
Now, when nsmux sets another dynamic translator, it sends a
retry to the client, so that it could set the (possible) next
dynamic translator in the next execution of netfs_S_dir_lookup,
which will occur due to the retry. However, nsmux cannot handle
the retry properly yet.

The reason for splitting the functionality of *only* passing the
retry to client in a separate commit is that it required some
tricky code manipulation and hence I feel that keeping it
in a separate commit is quite reasonable.