Adapted node_set_translator to setting translators on shadow nodes
Usually the file of which the supplied node is a mirror is
looked up by node_set_translator, so that it can open using
the flag combination required by the translator being started.
However, when dealing with a dynamic translator being set upon
another dynamic translator, node_set_translator must not do any
lookups, because a valid port is already stored in the supplied
node.