Dynamic translator stacks are set in series of retries
When a client wants nsmux to set several translators on a node
using a syntax like ``file,,x,,y'', nsmux sets the first
translator on a shadow node mirroring ``file'' and punts a
retry to the client. The client (commonly) does the retry and
thus control comes back to nsmux, which creates another shadow
node and sets the second translator on this node.

At the moment translator stacks do not include proxy nodes, i.e.
the structure of a dynamic translator stack is:

...---dynamic translator -- shadow node -- dynamic translator ---...