Added proxy nodes in the dynamic translator stack
When setting dynamic translators, nsmux used to offer the client
a port to the dynamic translator directly. Now it gives off a
proxy node, which contains a port to the translator. Also, proxy
nodes are now included in dynamic translator stacks. Dynamic
translator stacks now consist of the following interconnected
conceptual blocks:

  -- shadow node -- translator -- proxy node --