Shadow and proxy nodes are connected in a chain
nsmux will eventually need to be able to traverse dynamic
translator stacks, so it will have to know which node lies under
a certain node (shadow or proxy). To be able to do this, nsmux
now maintains the service nodes in a dynamic translator stack
in a one-directional linked list.